Some code cleanup.
