+++ UpkManager.Entities/UpkFile.cs
@@ -22,20 +22,4 @@ namespace UpkManager.Entities {
 
   }
 
-  public class ExportVersion {
-
-    public string Version { get; set; }
-
-    public List<ExportType> Types { get; set; }
-
-  }
-
-  public class ExportType {
-
-    public string Name { get; set; }
-
-    public List<string> ExportNames { get; set; }
-
-  }
-
 }
