+++ UpkManager.Entities/UpkManager.Entities.csproj
@@ -44,6 +44,8 @@
     <Reference Include="System.Core" />
   </ItemGroup>
   <ItemGroup>
+    <Compile Include="ExportType.cs" />
+    <Compile Include="ExportVersion.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="UpkFile.cs" />
     <Compile Include="UpkManagerException.cs" />
